+++ docs/CHANGELOG.md
@@ -85,6 +85,7 @@ All notable changes to this project are documented here.
 - Improved Swagger documentation using `ParameterObject`
 - Added API endpoint to cancel reservations (`PATCH /api/reservations/{id}/cancel`)
 - Added API endpoint to retrieve all reservations (`GET /api/reservations`)
+- Introduced ReservationDTO and updated reservation API to return DTO instead of entity.
 
 ---